+++ spec/support/puffing_billy.rb
@@ -5,34 +5,35 @@
 # 
 # 2. Add puffing-billy to Gemfile:
 #
-# group :test do
+# group :development, :test do
 #   gem 'puffing-billy'
 # end
 # 
 # 3. Create a file like this one you're reading in spec/support/puffing_billy.rb:
 require 'billy/rspec'
 
+# 4. Configure cache to behave as required. See all available options at:
+#    https://github.com/oesmith/puffing-billy#caching
 Billy.configure do |c|
   c.cache = true
   c.cache_request_headers = false
-  c.ignore_params = ["http://www.google-analytics.com/__utm.gif",
-                     "https://r.twimg.com/jot",
-                     "http://p.twitter.com/t.gif",
-                     "http://p.twitter.com/f.gif",
-                     "http://www.facebook.com/plugins/like.php",
-                     "https://www.facebook.com/dialog/oauth",
-                     "http://cdn.api.twitter.com/1/urls/count.json"]
-  c.path_blacklist = []
-  c.merge_cached_responses_whitelist = []
   c.persist_cache = true
-  c.ignore_cache_port = true # defaults to true
   c.non_successful_cache_disabled = false
   c.non_successful_error_level = :warn
-  c.non_whitelisted_requests_disabled = false
-  c.cache_path = 'spec/req_cache/'
+
+  # cache_path is where responses from external URLs will be saved as YAML.
+  c.cache_path = "spec/support/req_cache/"  
+
+  # Avoid having tests dependent on external URLs.
+  #
+  # Only set non_whitelisted_requests_disabled **temporarily**
+  # to false when first recording a 3rd party interaction. After
+  # the recording has been stored to cache_path, then set
+  # non_whitelisted_requests_disabled back to true.
+  c.non_whitelisted_requests_disabled = true
 end
 
-# 4. Uncomment the *_billy driver for your desired browser:
+# 5. Uncomment the *_billy driver for your desired browser:
 # Capybara.javascript_driver = :selenium_billy # Uses Firefox
 # Capybara.javascript_driver = :webkit_billy
 # Capybara.javascript_driver = :poltergeist_billy
@@ -47,11 +48,12 @@ Capybara.register_driver :selenium_chrome_billy do |app|
 end
 Capybara.javascript_driver = :selenium_chrome_billy
 
-# 5. Start using Puffing Billy. See spec/features/share_page_spec.rb for an example.
-
+# 6. Start using Puffing Billy. See spec/features/share_page_spec.rb for an example,
+#    and find your cached responses in spec/support/req_cache
 
 # Suggested docs
 # --------------
 # https://github.com/oesmith/puffing-billy
 # https://github.com/oesmith/puffing-billy#rspec-usage
+# https://github.com/oesmith/puffing-billy#caching